Move external paramenters and resolved dependencies logic out of v2alpha3
This is a refactor in preparation for v2alpha4 which uses in its majority the same logic as v2alpha3.

Logici and tests moved to chains/formats/slsa/internal to be able to use it from different versions.

Current behavior shouldn't change with these changes.
