+++ pkg/chains/formats/slsa/internal/resolved_dependencies/resolved_dependencies.go
@@ -17,10 +17,19 @@ limitations under the License.
 package resolveddependencies
 
 import (
+	"context"
 	"encoding/json"
+	"fmt"
 
 	"github.com/in-toto/in-toto-golang/in_toto/slsa_provenance/common"
-	v1 "github.com/in-toto/in-toto-golang/in_toto/slsa_provenance/v1"
+	slsa "github.com/in-toto/in-toto-golang/in_toto/slsa_provenance/v1"
+	"github.com/tektoncd/chains/pkg/chains/formats/slsa/internal/material"
+	"github.com/tektoncd/chains/pkg/chains/formats/slsa/internal/slsaconfig"
+	"github.com/tektoncd/chains/pkg/chains/objects"
+	v1 "github.com/tektoncd/pipeline/pkg/apis/pipeline/v1"
+	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
+	"go.uber.org/zap"
+	"knative.dev/pkg/logging"
 )
 
 const (
@@ -34,13 +43,19 @@ const (
 	InputResultName = "inputs/result"
 	// PipelineResourceName is the name of the resolved dependency of pipeline resource.
 	PipelineResourceName = "pipelineResource"
+	// v1beta1SpecResourceLabel is the name of the label that contains information about TaskRun resources.
+	v1beta1SpecResourceLabel = "tekton.dev/v1beta1-spec-resources"
 )
 
+// used to toggle the fields in  see AddTektonTaskDescriptor
+// and AddSLSATaskDescriptor
+type addTaskDescriptorContent func(*objects.TaskRunObjectV1) (*slsa.ResourceDescriptor, error) //nolint:staticcheck
+
 // ConvertMaterialToResolvedDependency converts a SLSAv0.2 Material to a resolved dependency
-func ConvertMaterialsToResolvedDependencies(mats []common.ProvenanceMaterial, name string) []v1.ResourceDescriptor {
-	rds := []v1.ResourceDescriptor{}
+func ConvertMaterialsToResolvedDependencies(mats []common.ProvenanceMaterial, name string) []slsa.ResourceDescriptor {
+	rds := []slsa.ResourceDescriptor{}
 	for _, mat := range mats {
-		rd := v1.ResourceDescriptor{}
+		rd := slsa.ResourceDescriptor{}
 		rd.URI = mat.URI
 		rd.Digest = mat.Digest
 		if len(name) > 0 {
@@ -53,15 +68,15 @@ func ConvertMaterialsToResolvedDependencies(mats []common.ProvenanceMaterial, na
 
 // RemoveDuplicateResolvedDependencies removes duplicate resolved dependencies from the slice of resolved dependencies.
 // Original order of resolved dependencies is retained.
-func RemoveDuplicateResolvedDependencies(resolvedDependencies []v1.ResourceDescriptor) ([]v1.ResourceDescriptor, error) {
-	out := make([]v1.ResourceDescriptor, 0, len(resolvedDependencies))
+func RemoveDuplicateResolvedDependencies(resolvedDependencies []slsa.ResourceDescriptor) ([]slsa.ResourceDescriptor, error) {
+	out := make([]slsa.ResourceDescriptor, 0, len(resolvedDependencies))
 
 	// make map to store seen resolved dependencies
 	seen := map[string]bool{}
 	for _, resolvedDependency := range resolvedDependencies {
 		// Since resolvedDependencies contain names, we want to ignore those while checking for duplicates.
 		// Therefore, make a copy of the resolved dependency that only contains the uri and digest fields.
-		rDep := v1.ResourceDescriptor{}
+		rDep := slsa.ResourceDescriptor{}
 		rDep.URI = resolvedDependency.URI
 		rDep.Digest = resolvedDependency.Digest
 		// pipelinTasks store content with the slsa-tekton buildType
@@ -85,4 +100,201 @@ func RemoveDuplicateResolvedDependencies(resolvedDependencies []v1.ResourceDescr
 		out = append(out, resolvedDependency)
 	}
 	return out, nil
+}
+
+// AddTektonTaskDescriptor returns the more verbose resolved dependency content. this adds the name, uri, digest
+// and content if possible.
+func AddTektonTaskDescriptor(tr *objects.TaskRunObjectV1) (*slsa.ResourceDescriptor, error) { //nolint:staticcheck
+	rd := slsa.ResourceDescriptor{}
+	storedTr, err := json.Marshal(tr)
+	if err != nil {
+		return nil, err
+	}
+
+	rd.Name = PipelineTaskConfigName
+	rd.Content = storedTr
+	if tr.Status.Provenance != nil && tr.Status.Provenance.RefSource != nil {
+		rd.URI = tr.Status.Provenance.RefSource.URI
+		rd.Digest = tr.Status.Provenance.RefSource.Digest
+	}
+
+	return &rd, nil
+}
+
+// AddSLSATaskDescriptor resolves dependency content for the more generic slsa verifiers. just logs
+// the name, uri and digest.
+func AddSLSATaskDescriptor(tr *objects.TaskRunObjectV1) (*slsa.ResourceDescriptor, error) { //nolint:staticcheck
+	if tr.Status.Provenance != nil && tr.Status.Provenance.RefSource != nil {
+		return &slsa.ResourceDescriptor{
+			Name:   PipelineTaskConfigName,
+			URI:    tr.Status.Provenance.RefSource.URI,
+			Digest: tr.Status.Provenance.RefSource.Digest,
+		}, nil
+	}
+	return nil, nil
+}
+
+// fromPipelineTask adds the resolved dependencies from pipeline tasks
+// such as pipeline task uri/digest for remote pipeline tasks and step and sidecar images.
+func fromPipelineTask(logger *zap.SugaredLogger, pro *objects.PipelineRunObjectV1, addTasks addTaskDescriptorContent) ([]slsa.ResourceDescriptor, error) {
+	pSpec := pro.Status.PipelineSpec
+	resolvedDependencies := []slsa.ResourceDescriptor{}
+	if pSpec != nil {
+		pipelineTasks := pSpec.Tasks
+		pipelineTasks = append(pipelineTasks, pSpec.Finally...)
+		for _, t := range pipelineTasks {
+			tr := pro.GetTaskRunFromTask(t.Name)
+			// Ignore Tasks that did not execute during the PipelineRun.
+			if tr == nil || tr.Status.CompletionTime == nil {
+				logger.Infof("taskrun status not found for task %s", t.Name)
+				continue
+			}
+			rd, err := addTasks(tr)
+			if err != nil {
+				logger.Errorf("error storing taskRun %s, error: %s", t.Name, err)
+				continue
+			}
+
+			if rd != nil {
+				resolvedDependencies = append(resolvedDependencies, *rd)
+			}
+
+			mats := []common.ProvenanceMaterial{}
+
+			// add step images
+			stepMaterials, err := material.FromStepImages(tr)
+			if err != nil {
+				return nil, err
+			}
+			mats = append(mats, stepMaterials...)
+
+			// add sidecar images
+			sidecarMaterials, err := material.FromSidecarImages(tr)
+			if err != nil {
+				return nil, err
+			}
+			mats = append(mats, sidecarMaterials...)
+
+			// convert materials to resolved dependencies
+			resolvedDependencies = append(resolvedDependencies, ConvertMaterialsToResolvedDependencies(mats, "")...)
+		}
+	}
+	return resolvedDependencies, nil
+}
+
+// taskDependencies gather all dependencies in a task and adds them to resolvedDependencies
+func taskDependencies(ctx context.Context, tro *objects.TaskRunObjectV1) ([]slsa.ResourceDescriptor, error) {
+	var resolvedDependencies []slsa.ResourceDescriptor
+	var err error
+	mats := []common.ProvenanceMaterial{}
+
+	// add step and sidecar images
+	stepMaterials, err := material.FromStepImages(tro)
+	if err != nil {
+		return nil, err
+	}
+	mats = append(mats, stepMaterials...)
+
+	sidecarMaterials, err := material.FromSidecarImages(tro)
+	if err != nil {
+		return nil, err
+	}
+	mats = append(mats, sidecarMaterials...)
+	resolvedDependencies = append(resolvedDependencies, ConvertMaterialsToResolvedDependencies(mats, "")...)
+
+	mats = material.FromTaskParamsAndResults(ctx, tro)
+	// convert materials to resolved dependencies
+	resolvedDependencies = append(resolvedDependencies, ConvertMaterialsToResolvedDependencies(mats, InputResultName)...)
+
+	// add task resources
+	// =====
+	// convert to v1beta1 and add any task resources
+	serializedResources := tro.Annotations[v1beta1SpecResourceLabel]
+	var resources v1beta1.TaskRunResources //nolint:staticcheck
+	shouldReplace := false
+	if err := json.Unmarshal([]byte(serializedResources), &resources); err == nil {
+		shouldReplace = true
+
+	}
+	trV1Beta1 := &v1beta1.TaskRun{} //nolint:staticcheck
+	fmt.Printf("%v", tro.GetObject().(*v1.TaskRun))
+	if err := trV1Beta1.ConvertFrom(ctx, tro.GetObject().(*v1.TaskRun)); err == nil {
+		if shouldReplace {
+			trV1Beta1.Spec.Resources = &resources //nolint:staticcheck
+		}
+		mats = material.FromTaskResources(ctx, trV1Beta1)
+
+	}
+
+	// convert materials to resolved dependencies
+	resolvedDependencies = append(resolvedDependencies,
+		ConvertMaterialsToResolvedDependencies(mats, PipelineResourceName)...)
+
+	// remove duplicate resolved dependencies
+	resolvedDependencies, err = RemoveDuplicateResolvedDependencies(resolvedDependencies)
+	if err != nil {
+		return nil, err
+	}
+
+	return resolvedDependencies, nil
+}
+
+// TaskRun constructs `predicate.resolvedDependencies` section by collecting all the artifacts that influence a taskrun such as source code repo and step&sidecar base images.
+func TaskRun(ctx context.Context, tro *objects.TaskRunObjectV1) ([]slsa.ResourceDescriptor, error) {
+	var resolvedDependencies []slsa.ResourceDescriptor
+	var err error
+
+	// add top level task config
+	if p := tro.Status.Provenance; p != nil && p.RefSource != nil {
+		rd := slsa.ResourceDescriptor{
+			Name:   TaskConfigName,
+			URI:    p.RefSource.URI,
+			Digest: p.RefSource.Digest,
+		}
+		resolvedDependencies = append(resolvedDependencies, rd)
+	}
+
+	rds, err := taskDependencies(ctx, tro)
+	if err != nil {
+		return nil, err
+	}
+	resolvedDependencies = append(resolvedDependencies, rds...)
+
+	return resolvedDependencies, nil
+}
+
+// PipelineRun constructs `predicate.resolvedDependencies` section by collecting all the artifacts that influence a pipeline run such as source code repo and step&sidecar base images.
+func PipelineRun(ctx context.Context, pro *objects.PipelineRunObjectV1, slsaconfig *slsaconfig.SlsaConfig, addTasks addTaskDescriptorContent) ([]slsa.ResourceDescriptor, error) {
+	var err error
+	var resolvedDependencies []slsa.ResourceDescriptor
+	logger := logging.FromContext(ctx)
+
+	// add pipeline config to resolved dependencies
+	if p := pro.Status.Provenance; p != nil && p.RefSource != nil {
+		rd := slsa.ResourceDescriptor{
+			Name:   PipelineConfigName,
+			URI:    p.RefSource.URI,
+			Digest: p.RefSource.Digest,
+		}
+		resolvedDependencies = append(resolvedDependencies, rd)
+	}
+
+	// add resolved dependencies from pipeline tasks
+	rds, err := fromPipelineTask(logger, pro, addTasks)
+	if err != nil {
+		return nil, err
+	}
+	resolvedDependencies = append(resolvedDependencies, rds...)
+
+	// add resolved dependencies from pipeline results
+	mats := material.FromPipelineParamsAndResults(ctx, pro, slsaconfig)
+	// convert materials to resolved dependencies
+	resolvedDependencies = append(resolvedDependencies, ConvertMaterialsToResolvedDependencies(mats, InputResultName)...)
+
+	// remove duplicate resolved dependencies
+	resolvedDependencies, err = RemoveDuplicateResolvedDependencies(resolvedDependencies)
+	if err != nil {
+		return nil, err
+	}
+	return resolvedDependencies, nil
 }
